polishing code

+++ src/main/java/com/md/mic/controller/VoiceRoomMicController.java
@@ -1,6 +1,6 @@
 package com.md.mic.controller;
 
-import com.md.mic.model.User;
+import com.md.mic.exception.UserNotFoundException;
 import com.md.mic.pojos.*;
 import com.md.mic.service.MicApplyUserService;
 import com.md.mic.service.VoiceRoomMicService;
@@ -34,28 +34,32 @@ public class VoiceRoomMicController {
     }
 
     @PostMapping("/voice/room/{roomId}/mic/apply")
-    public AddMicApplyResponse addMicApply(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody AddMicApplyRequest request) {
-        AddMicApplyResponse response = new AddMicApplyResponse(Boolean.TRUE);
+    public AddMicApplyResponse addMicApply(@PathVariable("roomId") String roomId,
+            @RequestBody AddMicApplyRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
             throw new IllegalArgumentException("room is not be found!");
         }
-        response.setResult(micApplyUserService.addMicApply(user.getUid(), roomInfo, request));
-        return response;
+        Boolean result = micApplyUserService.addMicApply(user.getUid(), roomInfo, request);
+        return new AddMicApplyResponse(result);
     }
 
     @DeleteMapping("/voice/room/{roomId}/mic/apply")
-    public DeleteMicApplyResponse deleteMicApply(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId) {
-        DeleteMicApplyResponse response = new DeleteMicApplyResponse(Boolean.TRUE);
+    public DeleteMicApplyResponse deleteMicApply(@PathVariable("roomId") String roomId,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
             throw new IllegalArgumentException("room is not be found!");
         }
         micApplyUserService.deleteMicApply(user.getUid(), roomId);
-        return response;
+        return new DeleteMicApplyResponse(Boolean.TRUE);
     }
 
     @GetMapping("/voice/room/{roomId}/mic")
@@ -69,51 +73,64 @@ public class VoiceRoomMicController {
 
     //闭麦
     @PostMapping("/voice/room/{roomId}/mic/close")
-    public CloseMicResponse closeMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody CloseMicRequest request) {
-        CloseMicResponse response = new CloseMicResponse(Boolean.TRUE);
+    public CloseMicResponse closeMic(@PathVariable("roomId") String roomId,
+            @RequestBody CloseMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
             throw new IllegalArgumentException("room is not be found!");
         }
         this.voiceRoomMicService.closeMic(user.getUid(), roomInfo.getChatroomId(), request.getIndex());
-        return response;
+        return new CloseMicResponse(Boolean.TRUE);
     }
 
     //取消关麦、开麦
     @DeleteMapping("/voice/room/{roomId}/mic/close")
-    public OpenMicResponse openMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody OpenMicRequest request) {
+    public OpenMicResponse openMic(@PathVariable("roomId") String roomId,
+            @RequestBody OpenMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         OpenMicResponse response = new OpenMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
             throw new IllegalArgumentException("room is not be found!");
         }
-        this.voiceRoomMicService.openMic(user.getUid(), roomInfo.getChatroomId(), request.getIndex());
+        this.voiceRoomMicService.openMic(user.getUid(), roomInfo.getChatroomId(),
+                request.getIndex());
         return response;
     }
 
     //下麦
     @DeleteMapping("/voice/room/{roomId}/mic/leave")
-    public LeaveMicResponse leaveMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody LeaveMicRequest request) {
+    public LeaveMicResponse leaveMic(@PathVariable("roomId") String roomId,
+            @RequestBody LeaveMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         LeaveMicResponse response = new LeaveMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
             throw new IllegalArgumentException("room is not be found!");
         }
-        this.voiceRoomMicService.leaveMic(user.getUid(), roomInfo.getChatroomId(), request.getIndex());
+        this.voiceRoomMicService.leaveMic(user.getUid(), roomInfo.getChatroomId(),
+                request.getIndex());
         return response;
     }
 
     //禁言麦位
     @PostMapping("/voice/room/{roomId}/mic/mute")
-    public MuteMicResponse muteMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody MuteMicRequest request) {
+    public MuteMicResponse muteMic(@PathVariable("roomId") String roomId,
+            @RequestBody MuteMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         MuteMicResponse response = new MuteMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
@@ -129,9 +146,12 @@ public class VoiceRoomMicController {
 
     //取消禁言
     @DeleteMapping("/voice/room/{roomId}/mic/mute")
-    public UnMuteMicResponse unMuteMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody UnMuteMicRequest request) {
+    public UnMuteMicResponse unMuteMic(@PathVariable("roomId") String roomId,
+            @RequestBody UnMuteMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         UnMuteMicResponse response = new UnMuteMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
@@ -147,23 +167,31 @@ public class VoiceRoomMicController {
 
     //交换麦位
     @PostMapping("/voice/room/{roomId}/mic/exchange")
-    public ExchangeMicResponse exchangeMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody ExchangeMicRequest request) {
+    public ExchangeMicResponse exchangeMic(@PathVariable("roomId") String roomId,
+            @RequestBody ExchangeMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         ExchangeMicResponse response = new ExchangeMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
             throw new IllegalArgumentException("room is not be found!");
         }
         this.voiceRoomMicService
-                .exchangeMic(roomInfo.getChatroomId(), request.getFrom(), request.getTo(), user.getUid());
+                .exchangeMic(roomInfo.getChatroomId(), request.getFrom(), request.getTo(),
+                        user.getUid());
         return response;
     }
 
     @PostMapping("/voice/room/{roomId}/mic/kick")
-    public KickUserMicResponse kickUserMic(@RequestAttribute("user") User user,
+    public KickUserMicResponse kickUserMic(
             @PathVariable("roomId") String roomId,
-            @RequestBody KickUserMicRequest request) {
+            @RequestBody KickUserMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         KickUserMicResponse response = new KickUserMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
@@ -172,16 +200,19 @@ public class VoiceRoomMicController {
         if (!roomInfo.getOwner().getUid().equals(user.getUid())) {
             throw new IllegalArgumentException("only the admin can kick mic");
         }
-        this.voiceRoomMicService.kickUserMic(roomInfo.getChatroomId(), request.getIndex(), request.getUid());
+        this.voiceRoomMicService.kickUserMic(roomInfo.getChatroomId(), request.getIndex(),
+                request.getUid());
 
         return response;
     }
 
     @PostMapping("/voice/room/{roomId}/mic/lock")
-    public LockMicResponse lockMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody LockMicRequest request) {
-
+    public LockMicResponse lockMic(@PathVariable("roomId") String roomId,
+            @RequestBody LockMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         LockMicResponse response = new LockMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
@@ -196,9 +227,12 @@ public class VoiceRoomMicController {
     }
 
     @DeleteMapping("/voice/room/{roomId}/mic/lock")
-    public UnLockMicResponse unLockMic(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody UnLockMicRequest request) {
+    public UnLockMicResponse unLockMic(@PathVariable("roomId") String roomId,
+            @RequestBody UnLockMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         UnLockMicResponse response = new UnLockMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
@@ -213,9 +247,12 @@ public class VoiceRoomMicController {
 
     //群主邀请上麦
     @PostMapping("/voice/room/{roomId}/mic/invite")
-    public InviteUserOnMicResponse invite(@RequestAttribute("user") User user,
-            @PathVariable("roomId") String roomId,
-            @RequestBody InviteUserOnMicRequest request) {
+    public InviteUserOnMicResponse invite(@PathVariable("roomId") String roomId,
+            @RequestBody InviteUserOnMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         InviteUserOnMicResponse response = new InviteUserOnMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
@@ -230,9 +267,13 @@ public class VoiceRoomMicController {
 
     //群主同意上麦申请
     @PostMapping("/voice/room/{roomId}/mic/apply/agree")
-    public ApplyAgreeOnMicResponse agreeApply(@RequestAttribute("user") User user,
+    public ApplyAgreeOnMicResponse agreeApply(
             @PathVariable("roomId") String roomId,
-            @RequestBody ApplyAgreeOnMicRequest request) {
+            @RequestBody ApplyAgreeOnMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         ApplyAgreeOnMicResponse response = new ApplyAgreeOnMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
@@ -247,17 +288,22 @@ public class VoiceRoomMicController {
     }
 
     //用户同意邀请上麦申请
-    @PostMapping("/voice/room/{roomId}/mic/invite/agree")
-    public InviteAgreeOnMicResponse agreeInvite(@RequestAttribute("user") User user,
+    @GetMapping("/voice/room/{roomId}/mic/invite/agree")
+    public InviteAgreeOnMicResponse agreeInvite(
             @PathVariable("roomId") String roomId,
-            @RequestBody InviteAgreeOnMicRequest request) {
+            @RequestBody InviteAgreeOnMicRequest request,
+            @RequestAttribute(name = "user", required = false) UserDTO user) {
+        if (user == null) {
+            throw new UserNotFoundException("user must not be null");
+        }
         InviteAgreeOnMicResponse response = new InviteAgreeOnMicResponse(Boolean.TRUE);
         VoiceRoomDTO roomInfo = voiceRoomService.getByRoomId(roomId);
         if (roomInfo == null) {
             throw new IllegalArgumentException("room is not be found!");
         }
         Boolean result =
-                voiceRoomMicService.agreeInvite(roomInfo.getChatroomId(), request.getUid(), request.getIndex());
+                voiceRoomMicService.agreeInvite(roomInfo.getChatroomId(), request.getUid(),
+                        request.getIndex());
         response.setResult(result);
         return response;
     }
